Remove the double delete command.

+++ netpyne/batch/batch.py
@@ -70,12 +70,13 @@ def runJob(script, cfgSavePath, netParamsSavePath):
 def createFolder(folder):
     import os
     if os.path.exists(folder):
-        try:
-            import shutil
-            shutil.rmtree(folder)
-            shutil.rmtree(folder)
-        except OSError:
-            print(' Could not delete %s' %(folder))
+        if pc.id()==0:
+            try:
+                import shutil
+                shutil.rmtree(folder)
+                # shutil.rmtree(folder)
+            except OSError:
+                print(' Could not delete %s' %(folder))
         
     if not os.path.exists(folder):
         try:
